ci/jenkins/scripts/python-ci-scripts/requirements.txt
requests==2.26.0
click==7.0
oyaml==1.0
yarl==1.8.1
aiohttp==3.8.3
docker==6.0.0
pytest==7.1.2
pytest-html==3.1.1
pylint==2.15.2
flake8==6.0.0
pydocstyle==6.3.0
semver==2.13.0
memoization==0.4.0
jira==3.5.2
markdown==3.4.4
yamllint==1.35.1
bs4==0.0.2
